+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.6.0 (unreleased)
 
 - rename `DB_URI` to `DATABASE_URI`
+- rename the `db_uri` argument of `init_sqlalchemy_unchained` to `database_uri`
 
 ## v0.5.1 (2018/10/20)